+++ gate/control/ControlModule.hpp
@@ -8,7 +8,14 @@
 #include <iomanip>
 #include <sstream>
 
-#include "codes/messages/led_set_intensity.hpp"
+#include <iostream>
+
+#include "codes/messages/led_panel/set_intensity.hpp"
+#include "codes/messages/led_panel/get_intensity_request.hpp"
+#include "codes/messages/led_panel/get_intensity_response.hpp"
+#include "codes/messages/led_panel/temperature_request.hpp"
+#include "codes/messages/led_panel/temperature_response.hpp"
+
 
 /**
  * @class ControlModule
@@ -26,6 +33,25 @@ public:
      * @param channel Channel to light up.
      */
     void setIntensity(Codes::Module module, float intensity, int channel, std::function<void(bool)> callback);
+
+    /**
+     * @brief Retrieves the current intensity of LEDs for a specified module and channel.
+     * 
+     * @param manager Reference to the CAN request manager, responsible for managing requests and responses.
+     * @param module Target module from which the LED intensity will be retrieved.
+     * @param channel The specific LED channel whose intensity will be retrieved. Valid channels range from 0 to 3.
+     */
+    void getIntensity(CanRequestManager& manager, Codes::Module module, int channel, std::function<void(float)> callback);
+
+    /**
+     * @brief Retrieves the current temperature of LEDs for a specified module and channel.
+     * 
+     * @param module Target module from which the LED temperature will be retrieved.
+     * @param callback Callback function to handle the retrieved temperature.
+     */
+    void getLedTemperature(CanRequestManager& manager, Codes::Module module, std::function<void(float)> callback); 
+
+
 
 protected:
     boost::asio::io_context& m_ioContext;