+++ oatpp-server/controller/MyController.cpp
@@ -24,7 +24,7 @@ std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::ge
 
     m_systemModule.getAvailableModules([&promise, dtoList](const std::vector<CanMessage>& responses) {
     for (const auto& response : responses) {
-        App_messages::Probe_modules_response moduleResponse;
+        App_messages::Common::Probe_modules_response moduleResponse;
 
         auto dataCopy = response.getData();
         if (moduleResponse.Interpret_data(dataCopy)) {
@@ -39,13 +39,13 @@ std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::ge
 
             uint32_t modulePart = (response.getId() >> 4) & 0xFF;
             switch (static_cast<Codes::Module>(modulePart)) {
-                case Codes::Module::Core_device:
+                case Codes::Module::Core_module:
                     moduleInfoDto->module_type = "core";
                     break;
-                case Codes::Module::Control_board:
+                case Codes::Module::Control_module:
                     moduleInfoDto->module_type = "control";
                     break;
-                case Codes::Module::Sensor_board:
+                case Codes::Module::Sensor_module:
                     moduleInfoDto->module_type = "sensor";
                     break;
                 default:
@@ -85,11 +85,11 @@ uint8_t MyController::getNextSeqNumber() {
 
 std::optional<Codes::Module> MyController::getTargetModule(const oatpp::Enum<dto::ModuleEnum>::AsString& module) {
     if (module == dto::ModuleEnum::control) {
-        return Codes::Module::Control_board;
+        return Codes::Module::Control_module;
     } else if (module == dto::ModuleEnum::sensor) {
-        return Codes::Module::Sensor_board;
+        return Codes::Module::Sensor_module;
     } else if (module == dto::ModuleEnum::core) {
-        return Codes::Module::Core_device;
+        return Codes::Module::Core_module;
     }
     return std::nullopt;
 }
@@ -126,7 +126,6 @@ std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::pi
     }
 }
 
-/*
 std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::getCoreLoad(const oatpp::Enum<dto::ModuleEnum>::AsString& module) {
     auto loadResponseDto = MyLoadResponseDto::createShared();
     std::promise<float> promise;
@@ -156,7 +155,6 @@ std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::ge
         return createResponse(Status::CODE_500, "Failed to retrieve load");
     }
 }
-*/
 
 std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::getCoreTemp(const oatpp::Enum<dto::ModuleEnum>::AsString& module) {
     auto tempResponseDto = MyTempDto::createShared();
@@ -206,7 +204,7 @@ std::future<bool> MyController::checkModuleAndUidAvailability(
         std::unordered_set<std::string> availableModules;
 
         for (const auto& response : responses) {
-            App_messages::Probe_modules_response moduleResponse;
+            App_messages::Common::Probe_modules_response moduleResponse;
             auto dataCopy = response.getData();
 
             if (moduleResponse.Interpret_data(dataCopy)) {
@@ -221,13 +219,13 @@ std::future<bool> MyController::checkModuleAndUidAvailability(
                 std::string moduleType;
 
                 switch (static_cast<Codes::Module>(modulePart)) {
-                    case Codes::Module::Core_device:
+                    case Codes::Module::Core_module:
                         moduleType = "core";
                         break;
-                    case Codes::Module::Control_board:
+                    case Codes::Module::Control_module:
                         moduleType = "control";
                         break;
-                    case Codes::Module::Sensor_board:
+                    case Codes::Module::Sensor_module:
                         moduleType = "sensor";
                         break;
                     default:
@@ -410,17 +408,79 @@ std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::po
 // ==========================================
 // Control module
 // ==========================================
-std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::setIntensity(const oatpp::Object<MyIntensityDto>& body) {
+std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::setIntensities(const oatpp::Object<MyIntensitiesDto>& body) {
+    if (!body || !body->intensity || body->intensity->size() != 4) {
+        return createResponse(Status::CODE_400, "Invalid intensity array. Must contain exactly 4 values.");
+    }
 
-   
+    for (size_t i = 0; i < body->intensity->size(); i++) {
+        auto intensity = body->intensity->at(i); 
+        if (!intensity || *intensity < 0.0f || *intensity > 1.0f) {
+            return createResponse(Status::CODE_400, "Invalid intensity value. Must be between 0.0 and 1.0.");
+        }
+    }
+
+    bool success = true;
+    for (size_t channel = 0; channel < body->intensity->size(); ++channel) {
+        std::promise<bool> promise;
+        auto future = promise.get_future();
+
+        auto handleSetIntensityResult = [&promise](bool result) {
+            promise.set_value(result);
+        };
+
+        m_controlModule.setIntensity(
+            Codes::Module::Control_module,
+            *body->intensity->at(channel), 
+            channel,
+            handleSetIntensityResult
+        );
+
+        future.wait();
+        if (!future.get()) {
+            success = false;
+            break;
+        }
+        std::this_thread::sleep_for(std::chrono::milliseconds(3));
+    }
+
+    if (success) {
+        return createResponse(Status::CODE_200, "Intensities set successfully.");
+    } else {
+        return createResponse(Status::CODE_500, "Failed to set intensities.");
+    }
+}
+
+std::optional<int> MyController::getTargetChannel(const dto::ChannelEnum& channel) {
+    switch (channel) {
+        case dto::ChannelEnum::channel0:
+            return 0;
+        case dto::ChannelEnum::channel1:
+            return 1;
+        case dto::ChannelEnum::channel2:
+            return 2;
+        case dto::ChannelEnum::channel3:
+            return 3;
+        default:
+            return std::nullopt;  
+    }
+}
+
+
+std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::setIntensity(
+    const oatpp::Enum<dto::ChannelEnum>::AsString& channel, 
+    const oatpp::Object<MyIntensityDto>& body
+) {
     if (body->intensity < 0 || body->intensity > 1) {
         return createResponse(Status::CODE_400, "Invalid intensity. Must be between 0 and 1.");
     }
 
-    
-    if (body->channel < 0 || body->channel > 3) { 
-        return createResponse(Status::CODE_400, "Invalid channel. Must be 0, 1, 2, or 3.");
+    auto targetChannelOpt = getTargetChannel(channel);
+    if (!targetChannelOpt.has_value()) { 
+        return createResponse(Status::CODE_400, "Invalid channel. Must be channel0, channel1, channel2, or channel3.");
     }
+
+    int targetChannel = targetChannelOpt.value();
 
     std::promise<bool> promise;
     auto future = promise.get_future();
@@ -429,8 +489,7 @@ std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::se
         promise.set_value(success);
     };
 
-    
-    m_controlModule.setIntensity(Codes::Module::Control_board, body->intensity, body->channel, handleSetIntensityResult);
+    m_controlModule.setIntensity(Codes::Module::Control_module, body->intensity, targetChannel, handleSetIntensityResult);
 
     future.wait();
     bool success = future.get();
@@ -441,6 +500,64 @@ std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::se
         return createResponse(Status::CODE_500, "Failed to set intensity.");
     }
 }
+
+std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::getIntensity(const oatpp::Enum<dto::ChannelEnum>::AsString& channel) {
+    auto intensityResponseDto = MyIntensityDto::createShared();
+    std::promise<float> promise;
+    auto future = promise.get_future();
+
+    auto handleIntensityResult = [&promise](float intensity) {
+        promise.set_value(intensity);
+    };
+
+    auto targetChannelOpt = getTargetChannel(channel);  
+    if (!targetChannelOpt.has_value()) {
+        return createResponse(Status::CODE_404, "Channel not found");
+    }
+    int targetChannel = targetChannelOpt.value();
+
+   
+    m_controlModule.getIntensity(m_canRequestManager, Codes::Module::Control_module, targetChannel, handleIntensityResult);
+    
+
+    future.wait();
+    float intensity = future.get();
+
+    if (intensity != -2 && intensity != -1) {
+        intensityResponseDto->intensity = intensity;
+        return createDtoResponse(Status::CODE_200, intensityResponseDto); 
+    } else if (intensity == -2) {
+        return createResponse(Status::CODE_504, "Request timed out"); 
+    } else {
+        return createResponse(Status::CODE_500, "Failed to retrieve intensity");  
+    }
+}
+
+std::shared_ptr<oatpp::web::protocol::http::outgoing::Response> MyController::getLedTemperature() {
+    auto tempResponseDto = MyTempDto::createShared();
+    std::promise<float> promise;
+    auto future = promise.get_future();
+
+    auto handleTempResult = [&promise](float temperature) {
+        promise.set_value(temperature);
+    };
+
+    m_controlModule.getLedTemperature(m_canRequestManager, Codes::Module::Control_module, handleTempResult);
+
+    future.wait();
+    float temperature = future.get();
+
+    if (temperature >= 0) {
+        tempResponseDto->temperature = temperature;
+        return createDtoResponse(Status::CODE_200, tempResponseDto);  
+    } else if (temperature == -2) {
+        return createResponse(Status::CODE_504, "Request timed out"); 
+    } else {
+        return createResponse(Status::CODE_500, "Failed to retrieve LED temperature");  
+    }
+}
+
+
 
 
 /*